Crate fj_interop

Fornjot Interop Types

Fornjot is an early-stage b-rep CAD kernel written in Rust. The kernel is split into multiple libraries that can be used semi-independently, and this is one of those.

This library defines types that allow other components of Fornjot to interoperate, without having to depend on each other.

Modules

ext

Extension traits for standard library types

Structs

CircleApproxParams

Approximation parameters for a circle

Color

RGBA color

InvalidTolerance

Error converting scalar to tolerance

MeshTriangle

A triangle in a TriMesh

Tolerance

A tolerance value

TriMesh

A triangle mesh

Functions

vertices_to_indexed_vertices

Convert a sequence of vertices into unique vertices and indices

Type Aliases

Index

An index that refers to a vertex in a mesh