fj_interop

Function vertices_to_indexed_vertices

pub fn vertices_to_indexed_vertices<I, O>(
    original_vertices: impl IntoIterator<Item = I>,
    map_vertex: impl Fn(I) -> O,
) -> (Vec<O>, Vec<Index>)
where
    I: Copy + Ord,

Convert a sequence of vertices into unique vertices and indices

Takes an iterator over vertices and converts those into two values:

A collection of vertices that contains all the original vertices, but each only once, with the order being undefined.

A collection of indices that correspond one-to-one to the original vertices reflect their original order.

The indices map into the collection of vertices.

Some environments, like renderers or some file formats, expect this kind of setup. These environments tend to also expect custom vertex formats, and those vertex formats might not fulfill the requirements needed to do the kind of comparing and sorting this function does.

For this reason, a closure that converts the original vertices into another format, after the comparing and sorting is done, is also accepted. This might save the caller from having to allocate another collection.